Extract PostHook to core and rename to EffectsGenerator.
