+++ flatland/envs/rail_env.py
@@ -8,6 +8,7 @@ from functools import lru_cache
 from typing import List, Optional, Dict, Tuple, Set
 
 import numpy as np
+from flatland.core.effects_creator import EffectsGenerator
 
 import flatland.envs.timetable_generators as ttg
 from flatland.core.env import Environment
@@ -25,8 +26,6 @@ from flatland.envs.agent_utils import EnvAgent
 from flatland.envs.distance_map import DistanceMap
 from flatland.envs.observations import GlobalObsForRailEnv
 from flatland.envs.rail_env_action import RailEnvActions
-from flatland.envs.rail_env_post_hook import PostHook
-from flatland.envs.step_utils import action_preprocessing
 from flatland.envs.rewards import Rewards
 from flatland.envs.step_utils import env_utils
 from flatland.envs.step_utils.action_preprocessing import process_illegal_action, check_valid_action, preprocess_left_right_action
@@ -92,7 +91,7 @@ class RailEnv(Environment):
                  acceleration_delta=1.0,
                  braking_delta=-1.0,
                  rewards: Rewards = None,
-                 post_hook: PostHook = None
+                 effects_generator: EffectsGenerator = None
                  ):
         """
         Environment init.
@@ -208,7 +207,7 @@ class RailEnv(Environment):
         self.acceleration_delta = acceleration_delta
         self.braking_delta = braking_delta
 
-        self.post_hook = post_hook
+        self.effects_generator = effects_generator
 
     def _seed(self, seed):
         self.np_random, seed = seeding.np_random(seed)
@@ -349,8 +348,8 @@ class RailEnv(Environment):
         self.agent_positions = np.zeros((self.height, self.width), dtype=int) - 1
         self._update_agent_positions_map(ignore_old_positions=False)
 
-        if self.post_hook is not None:
-            self.post_hook(self)
+        if self.effects_generator is not None:
+            self.effects_generator(self)
 
 
         self.dones = dict.fromkeys(list(range(self.get_num_agents())) + ["__all__"], False)
@@ -663,8 +662,8 @@ class RailEnv(Environment):
         if self.record_steps:
             self.record_timestep(action_dict)
 
-        if self.post_hook is not None:
-            self.post_hook(self)
+        if self.effects_generator is not None:
+            self.effects_generator(self)
 
         return self._get_observations(), self.rewards_dict, self.dones, self.get_info_dict()
